chore: download youtube-dl using ADD for better cache busting
I setup a cronjob to rebuild the container on Dockerhub each week,
but build caching was enabled which made the rebuilds pointless -
it was reusing cache and not redownloading youtube-dl.

I think ADD is always used regardless of cache. My intent is that,
if the youtube-dl binary has not been updated, cache can be used
and nothing needs to change. If the youtube-dl binary _has_ been
updated, the second stage should be cleanly built.